requirements.txt
Flask==3.0.0
google-generativeai==0.3.1
gunicorn==21.2.0
Flask-Session==0.5.0
python-dotenv==1.0.0
pymongo==4.5.0
